requirements.txt
# Core dependencies
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.4.0,<3.0.0
click>=8.1.0,<9.0.0

# Certificate handling
cryptography>=41.0.0,<42.0.0
pyOpenSSL>=23.3.0,<24.0.0

# Metrics and monitoring
prometheus-client>=0.19.0,<1.0.0
psutil>=5.9.0,<6.0.0

# Configuration and utilities
PyYAML>=6.0.1,<7.0.0
watchdog>=3.0.0,<4.0.0

# Async support
aiofiles>=23.2.1,<24.0.0